requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
async-timeout==5.0.1
asyncpg==0.30.0
certifi==2024.8.30
click==8.1.7
colorama==0.4.6
deprecated==1.2.15
dnspython==2.7.0
email-validator==2.2.0
et-xmlfile==2.0.0
exceptiongroup==1.2.2
fastapi==0.115.5
fastapi-cache2==0.2.2
fastapi-cli==0.0.5
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.0
idna==3.10
importlib-metadata==8.5.0
iniconfig==2.0.0
jinja2==3.1.4
limits==3.14.1
loguru==0.7.2
markdown-it-py==3.0.0
markupsafe==3.0.2
mdurl==0.1.2
numpy==2.1.3
openpyxl==3.1.5
opentelemetry-api==1.28.2
opentelemetry-instrumentation==0.49b2
opentelemetry-instrumentation-asgi==0.49b2
opentelemetry-instrumentation-fastapi==0.49b2
opentelemetry-sdk==1.28.2
opentelemetry-semantic-conventions==0.49b2
opentelemetry-util-http==0.49b2
packaging==24.2
pandas==2.2.3
pendulum==3.0.0
pluggy==1.5.0
psycopg2-binary==2.9.10
pydantic==2.10.2
pydantic-core==2.27.1
pydantic-settings==2.6.1
pygments==2.18.0
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.19
pytz==2024.2
pyyaml==6.0.2
redis==4.6.0
rich==13.9.4
shellingham==1.5.4
six==1.16.0
slowapi==0.1.9
sniffio==1.3.1
sqlalchemy==2.0.36
sqlmodel==0.0.22
starlette==0.41.3
time-machine==2.16.0
tomli==2.2.1
typer==0.14.0
typing-extensions==4.12.2
tzdata==2024.2
uvicorn==0.32.1
watchfiles==1.0.0
websockets==14.1
win32-setctime==1.1.0
wrapt==1.17.0
zipp==3.21.0
